Provide an option to disable grid animations
I was consistenly seeing flashes of white (iPhone X, iOS 11) when
transitioning between the grid and the photos view.

This commit provides a switch to turn off the transitions
animations. With this set to true, I no longer observe these flashes
in my project. However, this switch is by default set to false, so no
existing code will be impacted by this change.